Refactor server into modular packages
