+++ server/camera/__init__.py
@@ -1,0 +1,5 @@
+"""Camera-based gesture detection."""
+
+from . import dash_detection
+
+__all__ = ["dash_detection"]
