--- server/camera_recognizer/main.py
@@ -1,63 +1,0 @@
-import cv2
-import mediapipe as mp
-import numpy as np
-
-mp_pose = mp.solutions.pose
-pose = mp_pose.Pose()
-mp_draw = mp.solutions.drawing_utils
-
-def get_angle(a, b, c):
-    """Calculate the angle between 3 points."""
-    a = np.array(a)  # First
-    b = np.array(b)  # Mid
-    c = np.array(c)  # End
-    radians = np.arctan2(c[1]-b[1], c[0]-b[0]) - np.arctan2(a[1]-b[1], a[0]-b[0])
-    angle = np.abs(radians*180.0/np.pi)
-    return angle if angle < 180 else 360 - angle
-
-cap = cv2.VideoCapture(0)
-
-while cap.isOpened():
-    ret, frame = cap.read()
-    if not ret:
-        break
-
-    # Flip for natural interaction
-    frame = cv2.flip(frame, 1)
-
-    rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-    results = pose.process(rgb)
-
-    label = "Standing"
-
-    if results.pose_landmarks:
-        landmarks = results.pose_landmarks.landmark
-
-        # Get key points
-        hip = landmarks[mp_pose.PoseLandmark.LEFT_HIP]
-        knee = landmarks[mp_pose.PoseLandmark.LEFT_KNEE]
-        ankle = landmarks[mp_pose.PoseLandmark.LEFT_ANKLE]
-
-        # Get coordinates
-        hip_coords = [hip.x, hip.y]
-        knee_coords = [knee.x, knee.y]
-        ankle_coords = [ankle.x, ankle.y]
-
-        # Calculate knee angle
-        angle = get_angle(hip_coords, knee_coords, ankle_coords)
-
-        # Estimate crouch
-        if angle < 120:  # You can calibrate this threshold
-            label = "Crouching"
-
-        mp_draw.draw_landmarks(frame, results.pose_landmarks, mp_pose.POSE_CONNECTIONS)
-
-    # Display label
-    cv2.putText(frame, label, (50, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.2, (0, 255, 0), 3)
-    cv2.imshow("Posture Detection", frame)
-
-    if cv2.waitKey(10) & 0xFF == ord('q'):
-        break
-
-cap.release()
-cv2.destroyAllWindows()
