+++ server/core/__init__.py
@@ -1,0 +1,13 @@
+"""Core infrastructure for the motion controller server."""
+
+from .action_bus import ActionBus
+from .dispatcher import SensorDispatcher
+from .sensor_receiver import SensorReceiver
+from .input_handler import InputHandler
+
+__all__ = [
+    "ActionBus",
+    "SensorDispatcher",
+    "SensorReceiver",
+    "InputHandler",
+]
