+++ server/main.py
@@ -2,38 +2,14 @@
 import threading
 import time
 
-from dispatcher import SensorDispatcher
-from sensor_receiver import SensorReceiver
-from input_handler import InputHandler
-from motion_queue_recognizer import recognize as mqr_recognize
-
-# Optional dict-based recognizer
-try:
-    from recognize_movement import recognize_movement as rm_recognize
-    USE_RM = True
-except Exception:
-    USE_RM = False
-
-# Camera module we just wrapped
-import camera_recognizer.dash_detection as dash_detection
+from core import SensorDispatcher, SensorReceiver, InputHandler, ActionBus
+from phone import recognize as phone_recognize
+from camera import dash_detection
 
 from config import CAMERA_ACTION_MAP  # e.g. {"dash_left":"dash_out", "dash_right":"dash_in", "jump":"jump"}
 
-# ActionBus for volume keys
-try:
-    from action_bus import ActionBus
-    HAVE_ACTIONBUS = True
-except Exception:
-    HAVE_ACTIONBUS = False
-
 ih = InputHandler()
 disp = SensorDispatcher()
-
-# Keep last known sensor values for rm_recognize(data_dict)
-last_sample = {
-    "accelerometer": {"x": 0.0, "y": 0.0, "z": 0.0},
-    "gyroscope": {"x": 0.0, "y": 0.0, "z": 0.0},
-}
 
 def recognizer_callback(sensor_type: str, parts):
     """
@@ -45,24 +21,13 @@ def recognizer_callback(sensor_type: str, parts):
         if len(parts) >= 4:
             x, y, z = float(parts[1]), float(parts[2]), float(parts[3])
 
-        # Update dict for rm_recognize
-        if sensor_type == "accel":
-            last_sample["accelerometer"] = {"x": x, "y": y, "z": z}
-        elif sensor_type == "gyro":
-            last_sample["gyroscope"] = {"x": x, "y": y, "z": z}
-
-        # (1) queue-based recognizer
-        act1 = mqr_recognize(sensor_type, x, y, z)
-        # (2) dict-based recognizer
-        act2 = rm_recognize(last_sample) if USE_RM else None
-
-        action_raw = act1 or act2
+        action_raw = phone_recognize(sensor_type, x, y, z)
         if not action_raw:
             return
 
-        # Normalize action names (java_server uses "Attack", "Parry", "Dash in/out", ...)
+        # Normalize action names (e.g., "Attack", "Parry", "Dash in/out")
         normalized = action_raw.lower().replace(" ", "_")
-        # ih.handle(normalized)
+        ih.handle(normalized)
 
     except Exception as e:
         print(f"[main] recognizer_callback error: {e} | parts={parts}")
@@ -87,8 +52,6 @@ def start_camera():
     dash_detection.run(on_event=on_cam_event, show_window=True)
 
 def _wire_actionbus():
-    if not HAVE_ACTIONBUS:
-        return
     # Your phone sends numeric volkey codes; map them here
     def _on_vk(code: str, ts_ms: int):
         if str(code) == "25":   # volume down
