+++ server/phone/__init__.py
@@ -1,0 +1,5 @@
+"""Phone-based motion recognizers."""
+
+from .queue_recognizer import recognize
+
+__all__ = ["recognize"]
